add. move DNS functions into a mixin
