+++ lib/msf/core/auxiliary/mixins.rb
@@ -35,6 +35,7 @@ require 'msf/core/auxiliary/pii'
 require 'msf/core/auxiliary/redis'
 require 'msf/core/auxiliary/sms'
 require 'msf/core/auxiliary/mms'
+require 'msf/core/auxiliary/dns'
 
 #
 # Custom HTTP modules
